Replace custom serialization methods with pickle
The goal of this change is to reduce the maintenance overhead and code
complexity by removing the custom serialization.

This breaks python2 compatibility, which is okay because this is a part of
several changes - one of which is to remove python2 support entirely.

Instead of serialising everything to json the standard pickle module is used.
Because pickle is a standard library we don't need to test it!

Having a standard text format for output is an important feature; this
will be added back in another commit.

Fix runtime error from dodgy rebase