Add --binary option to the dexter test tool
This option allows a user to specify a prebuilt binary to pass to the debugger.

Either or both of --builder and --binary options are required. --binary always
overrides --builder if both are given.

The test directory option is still required so that dexter can read in the
dexter commands (DexWatch etc).

This doesn't work with multiple tests as only one binary can be specified.
Ignoring this for now because the ability to run multiple tests will be
removed soon (in favour of using llvm-lit framework).